+++ ja2lib/Tactical/Interface.c
@@ -606,7 +606,7 @@ void PopupMovementMenu(UI_EVENT *pUIEvent) {
   }
   SetButtonFastHelpText(iActionIcons[RUN_ICON], pTacticalPopupButtonStrings[RUN_ICON]);
   // SetButtonSavedRect( iActionIcons[ RUN_ICON ] );
-  ButtonList[iActionIcons[RUN_ICON]]->UserData[0] = (UINT32)pUIEvent;
+  ButtonList[iActionIcons[RUN_ICON]]->UserData[0] = (uintptr_t)pUIEvent;
 
   if (MercInWater(pSoldier) || (pSoldier->uiStatusFlags & SOLDIER_VEHICLE) ||
       (pSoldier->uiStatusFlags & SOLDIER_ROBOT)) {
@@ -628,7 +628,7 @@ void PopupMovementMenu(UI_EVENT *pUIEvent) {
     SetButtonFastHelpText(iActionIcons[WALK_ICON], pTacticalPopupButtonStrings[WALK_ICON]);
   }
 
-  ButtonList[iActionIcons[WALK_ICON]]->UserData[0] = (UINT32)pUIEvent;
+  ButtonList[iActionIcons[WALK_ICON]]->UserData[0] = (uintptr_t)pUIEvent;
 
   if (pSoldier->uiStatusFlags & SOLDIER_ROBOT) {
     if (!CanRobotBeControlled(pSoldier)) {
@@ -646,7 +646,7 @@ void PopupMovementMenu(UI_EVENT *pUIEvent) {
   }
   SetButtonFastHelpText(iActionIcons[SNEAK_ICON], pTacticalPopupButtonStrings[SNEAK_ICON]);
   // SetButtonSavedRect( iActionIcons[ SNEAK_ICON ] );
-  ButtonList[iActionIcons[SNEAK_ICON]]->UserData[0] = (UINT32)pUIEvent;
+  ButtonList[iActionIcons[SNEAK_ICON]]->UserData[0] = (uintptr_t)pUIEvent;
 
   // Check if this is a valid stance, diable if not!
   if (!IsValidStance(pSoldier, ANIM_CROUCH)) {
@@ -663,7 +663,7 @@ void PopupMovementMenu(UI_EVENT *pUIEvent) {
   }
   SetButtonFastHelpText(iActionIcons[CRAWL_ICON], pTacticalPopupButtonStrings[CRAWL_ICON]);
   // SetButtonSavedRect( iActionIcons[ CRAWL_ICON ] );
-  ButtonList[iActionIcons[CRAWL_ICON]]->UserData[0] = (UINT32)pUIEvent;
+  ButtonList[iActionIcons[CRAWL_ICON]]->UserData[0] = (uintptr_t)pUIEvent;
 
   // Check if this is a valid stance, diable if not!
   if (!IsValidStance(pSoldier, ANIM_PRONE)) {
@@ -679,7 +679,7 @@ void PopupMovementMenu(UI_EVENT *pUIEvent) {
   }
   SetButtonFastHelpText(iActionIcons[LOOK_ICON], TacticalStr[LOOK_CURSOR_POPUPTEXT]);
   // SetButtonSavedRect( iActionIcons[ LOOK_ICON ] );
-  ButtonList[iActionIcons[LOOK_ICON]]->UserData[0] = (UINT32)pUIEvent;
+  ButtonList[iActionIcons[LOOK_ICON]]->UserData[0] = (uintptr_t)pUIEvent;
 
   if (pSoldier->uiStatusFlags & SOLDIER_VEHICLE) {
     DisableButton(iActionIcons[LOOK_ICON]);
@@ -761,7 +761,7 @@ void PopupMovementMenu(UI_EVENT *pUIEvent) {
   }
   // SetButtonSavedRect( iActionIcons[ ACTIONC_ICON ] );
   SetButtonFastHelpText(iActionIcons[ACTIONC_ICON], zActionString);
-  ButtonList[iActionIcons[ACTIONC_ICON]]->UserData[0] = (UINT32)pUIEvent;
+  ButtonList[iActionIcons[ACTIONC_ICON]]->UserData[0] = (uintptr_t)pUIEvent;
 
   if (fDisableAction) {
     DisableButton(iActionIcons[ACTIONC_ICON]);
@@ -776,7 +776,7 @@ void PopupMovementMenu(UI_EVENT *pUIEvent) {
   }
   // SetButtonSavedRect( iActionIcons[ TALK_ICON ] );
   SetButtonFastHelpText(iActionIcons[TALK_ICON], pTacticalPopupButtonStrings[TALK_ICON]);
-  ButtonList[iActionIcons[TALK_ICON]]->UserData[0] = (UINT32)pUIEvent;
+  ButtonList[iActionIcons[TALK_ICON]]->UserData[0] = (uintptr_t)pUIEvent;
 
   if (AM_AN_EPC(pSoldier) || (pSoldier->uiStatusFlags & SOLDIER_VEHICLE)) {
     DisableButton(iActionIcons[TALK_ICON]);
@@ -792,7 +792,7 @@ void PopupMovementMenu(UI_EVENT *pUIEvent) {
   }
   // SetButtonSavedRect( iActionIcons[ HAND_ICON ] );
   SetButtonFastHelpText(iActionIcons[HAND_ICON], pTacticalPopupButtonStrings[HAND_ICON]);
-  ButtonList[iActionIcons[HAND_ICON]]->UserData[0] = (UINT32)pUIEvent;
+  ButtonList[iActionIcons[HAND_ICON]]->UserData[0] = (uintptr_t)pUIEvent;
 
   if (AM_AN_EPC(pSoldier) || (pSoldier->uiStatusFlags & SOLDIER_VEHICLE)) {
     DisableButton(iActionIcons[HAND_ICON]);
@@ -808,7 +808,7 @@ void PopupMovementMenu(UI_EVENT *pUIEvent) {
   }
   // SetButtonSavedRect( iActionIcons[ CANCEL_ICON ] );
   SetButtonFastHelpText(iActionIcons[CANCEL_ICON], pTacticalPopupButtonStrings[CANCEL_ICON]);
-  ButtonList[iActionIcons[CANCEL_ICON]]->UserData[0] = (UINT32)pUIEvent;
+  ButtonList[iActionIcons[CANCEL_ICON]]->UserData[0] = (uintptr_t)pUIEvent;
 
   // LockTacticalInterface( );
