+++ ja2lib/Tactical/InterfaceDialogue.h
@@ -44,7 +44,7 @@ typedef struct {
 extern NPC_DIALOGUE_TYPE gTalkPanel;
 
 BOOLEAN InitiateConversation(struct SOLDIERTYPE *pDestSoldier, struct SOLDIERTYPE *pSrcSoldier,
-                             INT8 bApproach, UINT32 uiApproachData);
+                             INT8 bApproach, uintptr_t uiApproachData);
 
 // THis fuction will allocate and setup an NPCDiaogue structure. Loads the face for the character..
 BOOLEAN InitTalkingMenu(UINT8 ubCharacterNum, INT16 sGridNo);
